gui start

+++ shared/bitflag/bitflag.hpp
@@ -12,7 +12,7 @@ namespace shared
 		/// <summary>
 		/// Creates a bitflag object with the given flags
 		/// </summary>
-		bitflag_t( uint32_t flags ) : m_flags( flags ) {}
+		bitflag_t( uintptr_t flags ) : m_flags( flags ) {}
 
 		/// <summary>
 		/// Destroys the bitflag object
@@ -22,14 +22,14 @@ namespace shared
 		/// <summary>
 		/// Stored flags
 		/// </summary>
-		uint32_t m_flags{};
+		uintptr_t m_flags{};
 
 		/// <summary>
 		/// Checks whether the given flag is present in the stored flags
 		/// </summary>
 		/// <param name="flag">Flag that will be checked</param>
 		/// <returns>Returns true if the flag is present in the stored flags</returns>
-		inline bool has_flag( uint32_t flag )
+		inline bool has_flag( uintptr_t flag )
 		{
 			return m_flags & flag;
 		}
@@ -38,7 +38,7 @@ namespace shared
 		/// Adds the given flag to the stored flags
 		/// </summary>
 		/// <param name="flag">Flag that will be added</param>
-		inline void add_flag( uint32_t flag )
+		inline void add_flag( uintptr_t flag )
 		{
 			m_flags |= flag;
 		}
@@ -47,7 +47,7 @@ namespace shared
 		/// Removes the given flag to the stored flags
 		/// </summary>
 		/// <param name="flag">Flag that will be removed</param>
-		inline void remove_flag( uint32_t flag )
+		inline void remove_flag( uintptr_t flag )
 		{
 			m_flags &= ~flag;
 		}
